backend/requirements.txt
Django==3.2.16
python-dotenv
psycopg2-binary==2.8.6
djangorestframework==3.12.4
Pillow==9.3.0
djoser==2.1.0
django-filter